fix: change price type from BigDecimal to Integer across product-related classes

+++ product-search-service/src/main/java/com/zwap/product_search_service/dto/ProductSearchQry.java
@@ -4,7 +4,6 @@ import com.zwap.product_search_service.enums.ProductSortEnum;
 import jakarta.validation.constraints.*;
 import lombok.Data;
 
-import java.math.BigDecimal;
 
 
 @Data
@@ -13,10 +12,10 @@ public class ProductSearchQry {
     private String searchParam;
     @Min(value = 0, message = "min price must be at least 0")
     @Max(value = 1000000, message = "min price must not exceed 1,000,000")
-    private BigDecimal minPrice;
+    private Integer minPrice;
     @Min(value = 0, message = "max price must be at least 0")
     @Max(value = 1000000, message = "max price must not exceed 1,000,000")
-    private BigDecimal maxPrice;
+    private Integer maxPrice;
     @Min(value = 0, message = "min distance must be at least 0")
     @Max(value = 1000000, message = "max distance must not exceed 1,000,000")
     private Integer distance;
